exercicios

+++ 035.py
@@ -1,6 +1,6 @@
-lin1 = int(input('Digite o comprimento da primeira reta: '))
-lin2 = int(input('Digite o comprimento da segunda reta: '))
-lin3 = int(input('Digite o comprimento da terceira reta: '))
+lin1 = float(input('Digite o comprimento da primeira reta: '))
+lin2 = float(input('Digite o comprimento da segunda reta: '))
+lin3 = float(input('Digite o comprimento da terceira reta: '))
 if lin1 < lin2 + lin3:
     if lin2 < lin1 + lin3:
         if lin3 < lin1 + lin2:
